pyproject.toml
[project]
name = "bbscan"
version = "0.1.0"
description = "Asynchronous bug bounty scanner"
authors = [{name="AutoGenerated", email="auto@example.com"}]
requires-python = ">=3.11"
dependencies = [
    "httpx[http2]",
    "pandas",
    "jsonschema",
    "typer",
    "structlog",
    "aiodns",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-asyncio",
    "vcrpy",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.typer]
main = "app.cli:app"

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true


[tool.setuptools]
packages = ["app"]

[project.scripts]
bbscan = "app.cli:app"
